backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
PyJWT==2.8.0 # For potential future token-based auth